+++ src/Entity/Tenant/Url.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity;
+namespace App\Entity\Tenant;
 
 use App\Repository\UrlRepository;
 use Doctrine\ORM\Mapping as ORM;
@@ -9,12 +9,8 @@ use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity(repositoryClass: UrlRepository::class)]
 #[ORM\HasLifecycleCallbacks]
-class Url
+class Url extends AbstractTenantScopedEntity
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
     #[Assert\NotBlank(message: new TranslatableMessage('The URL field cannot be empty.'))]
     #[Assert\Url(message: new TranslatableMessage('The value "{{ value }}" is not a valid URL.'))]
     #[ORM\Column(length: 255)]
@@ -23,11 +19,6 @@ class Url
     #[ORM\ManyToOne(targetEntity: Qr::class, inversedBy: 'urls')]
     #[ORM\JoinColumn(nullable: true)]
     private ?Qr $qr = null;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
 
     public function getUrl(): ?string
     {