+++ src/Utils/Roles.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Utils;
+
+class Roles
+{
+    final public const string ROLE_USER = 'ROLE_USER';
+    final public const string ROLE_EDITOR = 'ROLE_EDITOR';
+    final public const string ROLE_ADMIN = 'ROLE_ADMIN';
+    final public const string ROLE_SUPER_ADMIN = 'ROLE_SUPER_ADMIN';
+}